requirements.txt

google-generativeai
python-dotenv
pydantic
pillow
# Core dependencies

# AI metadata extraction
sd-parsers>=0.6.0

